Renamed namespace from "Application" to "WebUI"
