+++ Development/WebUI/Controllers/TranslationController.cs
@@ -6,7 +6,7 @@ using Service;
 using System;
 using System.Linq;
 
-namespace Application.Controllers
+namespace WebUI.Controllers
 {
     [ApiController]
     [Route("api/translate")]
