+++ Development/WebUI/Controllers/WordController.cs
@@ -8,7 +8,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Application.Controllers
+namespace WebUI.Controllers
 {
     [ApiController]
     [Route("api/word")]
