+++ Development/WebUI/Filters/ValidationActionFilter.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 
-namespace Application.Filters
+namespace WebUI.Filters
 {
     public class ValidationActionFilter : IActionFilter
     {
